+++ src/Opw.HttpExceptions/HttpExceptionBase.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Net;
 
 namespace Opw.HttpExceptions
@@ -12,6 +12,13 @@ namespace Opw.HttpExceptions
         /// HTTP status code.
         /// </summary>
         public abstract HttpStatusCode StatusCode { get; }
+
+        /// <summary>
+        /// Gets or sets a link to the help file associated with this exception.
+        /// For HttpExeptions a link to status code information https://tools.ietf.org/html/rfc7231.
+        /// </summary>
+        /// <returns>The Uniform Resource Name (URN) or Uniform Resource Locator (URL).</returns>
+        public abstract override string HelpLink { get; set; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="HttpExceptionBase"></see> class with status code InternalServerError.