+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-Change UseInvalidModelStateResponseFactory to SuppressInvalidModelStateResponseFactoryOverride
+Use rfc7231 docs as info for status codes (https://tools.ietf.org/html/rfc7231).
+
+As part of this release we had [1](https://github.com/ofpinewood/http-exceptions/milestone/4?closed=1) issues closed.
+
+## Bugs
+No bugs were fixed in this release.
 
 ## Improvements/Features
-* [??](https://github.com/ofpinewood/http-exceptions/issues/??) ?? 
+* [#17](https://github.com/ofpinewood/http-exceptions/issues/17) Use rfc7231 docs as info for status codes